requirements.txt
sentence-transformers
PyPDF2
python-docx
torch